move base CondaError to conda.__init__; register subclasses
