+++ conda/__init__.py
@@ -10,6 +10,7 @@ import os
 import sys
 
 from ._vendor.auxlib.packaging import get_version
+from .common.compat import with_metaclass
 
 __all__ = [
     "__name__", "__version__", "__author__",
@@ -27,3 +28,28 @@ __url__ = "https://github.com/conda/conda"
 
 if os.getenv('CONDA_ROOT') is None:
     os.environ['CONDA_ROOT'] = sys.prefix
+
+
+class CondaErrorType(type):
+
+    def __init__(cls, name, bases, attr):
+        super(CondaErrorType, cls).__init__(name, bases, attr)
+        key = "%s.%s" % (cls.__module__, name)
+        if key == "conda.CondaError":
+            cls.registry = dict()
+        else:
+            cls.registry[cls.__name__] = cls
+
+
+@with_metaclass(CondaErrorType)
+class CondaError(Exception):
+    def __init__(self, *args, **kwargs):
+        super(CondaError, self).__init__(*args, **kwargs)
+
+    def __repr__(self):
+        ret_str = ' '.join([str(arg) for arg in self.args if not isinstance(arg, bool)])
+        return ret_str
+
+    def __str__(self):
+        ret_str = ' '.join([str(arg) for arg in self.args if not isinstance(arg, bool)])
+        return ret_str
