+++ conda/cli/install.py
@@ -23,8 +23,9 @@ from ..cli.find_commands import find_executable
 from ..exceptions import (CondaFileNotFoundError, CondaValueError, DirectoryNotFoundError,
                           CondaEnvironmentError, PackageNotFoundError, TooManyArgumentsError,
                           CondaAssertionError, CondaOSError, CondaImportError,
-                          CondaError, DryRunExit, LockError, CondaRuntimeError,
+                          DryRunExit, LockError, CondaRuntimeError,
                           CondaSystemExit, NoPackagesFoundError, UnsatisfiableError, CondaIOError)
+from conda import CondaError
 from ..install import linked as install_linked
 from ..install import name_dist, is_linked
 from ..misc import explicit, clone_env, append_env, touch_nonadmin
