+++ tests/test_exceptions.py
@@ -1,6 +1,7 @@
 import random
 import unittest
 
+import conda
 from conda import exceptions
 
 
@@ -11,7 +12,7 @@ class InvalidInstructionTestCase(unittest.TestCase):
 
     def test_extends_from_conda_exception(self):
         e = exceptions.InvalidInstruction("foo")
-        self.assertIsInstance(e, exceptions.CondaError)
+        self.assertIsInstance(e, conda.CondaError)
 
     def test_creates_message_with_instruction_name(self):
         random_instruction = random.randint(100, 200)
@@ -22,26 +23,26 @@ class InvalidInstructionTestCase(unittest.TestCase):
 
 class CondaErrorTestCase(unittest.TestCase):
     def test_repr_is_correct(self):
-        e = exceptions.CondaError("Can't do that")
+        e = conda.CondaError("Can't do that")
         self.assertEqual(repr(e), "Can't do that")
 
     def test_inherited_repr_is_correct(self):
         try:
             raise exceptions.CondaValueError('value incorrect')
         except exceptions.CondaValueError as e:
-            err = exceptions.CondaError("Can't do that", e)
+            err = conda.CondaError("Can't do that", e)
 
         self.assertEqual(repr(err), "Can't do that Value error: value incorrect\n")
 
     def test_str_is_correct(self):
-        e = exceptions.CondaError("Can't do that")
+        e = conda.CondaError("Can't do that")
         self.assertEqual(str(e), "Can't do that")
 
     def test_inherited_str_is_correct(self):
         try:
             raise exceptions.CondaValueError('value incorrect')
         except exceptions.CondaValueError as e:
-            err = exceptions.CondaError("Can't do that", e)
+            err = conda.CondaError("Can't do that", e)
 
         self.assertEqual(str(err), "Can't do that Value error: value incorrect\n")
 
@@ -49,7 +50,7 @@ class CondaErrorTestCase(unittest.TestCase):
         try:
             raise exceptions.CondaValueError('value incorrect', False)
         except exceptions.CondaValueError as e:
-            err = exceptions.CondaError("Can't do that", e, True)
+            err = conda.CondaError("Can't do that", e, True)
 
         self.assertNotIn('False', repr(err))
         self.assertNotIn('True', repr(err))
